day 9: Sensor Boost: refactored Program to support relative addresses, big numbers, extra memory

+++ src/test/java/day7/PhaseSettingsTest.java
@@ -13,14 +13,14 @@ class PhaseSettingsTest {
     @Test
     void testCase1() {
         int[] program = {3, 15, 3, 16, 1002, 16, 10, 16, 1, 16, 15, 15, 4, 15, 99, 0, 0};
-        int output = PhaseSettings.calculateThrusterOutput(program, List.of(4, 3, 2, 1, 0));
+        long output = PhaseSettings.calculateThrusterOutput(program, List.of(4, 3, 2, 1, 0));
         assertThat(output).isEqualTo(43210);
     }
 
     @Test
     void testCase1MaxOutput() {
         int[] program = {3, 15, 3, 16, 1002, 16, 10, 16, 1, 16, 15, 15, 4, 15, 99, 0, 0};
-        int output = PhaseSettings.calculateMaxThrusterOutput(program);
+        long output = PhaseSettings.calculateMaxThrusterOutput(program);
         assertThat(output).isEqualTo(43210);
     }
 
@@ -28,7 +28,7 @@ class PhaseSettingsTest {
     void testCase2() {
         int[] program = {3, 23, 3, 24, 1002, 24, 10, 24, 1002, 23, -1, 23,
                 101, 5, 23, 23, 1, 24, 23, 23, 4, 23, 99, 0, 0};
-        int output = PhaseSettings.calculateThrusterOutput(program, List.of(0, 1, 2, 3, 4));
+        long output = PhaseSettings.calculateThrusterOutput(program, List.of(0, 1, 2, 3, 4));
         assertThat(output).isEqualTo(54321);
     }
 
@@ -36,7 +36,7 @@ class PhaseSettingsTest {
     void testCase2MaxOutput() {
         int[] program = {3, 23, 3, 24, 1002, 24, 10, 24, 1002, 23, -1, 23,
                 101, 5, 23, 23, 1, 24, 23, 23, 4, 23, 99, 0, 0};
-        int output = PhaseSettings.calculateMaxThrusterOutput(program);
+        long output = PhaseSettings.calculateMaxThrusterOutput(program);
         assertThat(output).isEqualTo(54321);
     }
 
@@ -44,7 +44,7 @@ class PhaseSettingsTest {
     void testCase3() {
         int[] program = {3, 31, 3, 32, 1002, 32, 10, 32, 1001, 31, -2, 31, 1007, 31, 0, 33,
                 1002, 33, 7, 33, 1, 33, 31, 31, 1, 32, 31, 31, 4, 31, 99, 0, 0, 0};
-        int output = PhaseSettings.calculateThrusterOutput(program, List.of(1, 0, 4, 3, 2));
+        long output = PhaseSettings.calculateThrusterOutput(program, List.of(1, 0, 4, 3, 2));
         assertThat(output).isEqualTo(65210);
     }
 
@@ -52,7 +52,7 @@ class PhaseSettingsTest {
     void testCase3MaxOutput() {
         int[] program = {3, 31, 3, 32, 1002, 32, 10, 32, 1001, 31, -2, 31, 1007, 31, 0, 33,
                 1002, 33, 7, 33, 1, 33, 31, 31, 1, 32, 31, 31, 4, 31, 99, 0, 0, 0};
-        int output = PhaseSettings.calculateMaxThrusterOutput(program);
+        long output = PhaseSettings.calculateMaxThrusterOutput(program);
         assertThat(output).isEqualTo(65210);
     }
 
@@ -60,7 +60,7 @@ class PhaseSettingsTest {
     void feedbackLoopTestCase1() {
         int[] program = {3, 26, 1001, 26, -4, 26, 3, 27, 1002, 27, 2, 27, 1, 27, 26,
                 27, 4, 27, 1001, 28, -1, 28, 1005, 28, 6, 99, 0, 0, 5};
-        int output = PhaseSettings.calculateFeedbackThrusterOutput(program, List.of(9, 8, 7, 6, 5));
+        long output = PhaseSettings.calculateFeedbackThrusterOutput(program, List.of(9, 8, 7, 6, 5));
         assertThat(output).isEqualTo(139629729);
     }
 
@@ -68,7 +68,7 @@ class PhaseSettingsTest {
     void feedbackLoopTestCase1MaxOutput() {
         int[] program = {3, 26, 1001, 26, -4, 26, 3, 27, 1002, 27, 2, 27, 1, 27, 26,
                 27, 4, 27, 1001, 28, -1, 28, 1005, 28, 6, 99, 0, 0, 5};
-        int output = PhaseSettings.calculateFeedbackThrusterMaxOutput(program);
+        long output = PhaseSettings.calculateFeedbackThrusterMaxOutput(program);
         assertThat(output).isEqualTo(139629729);
     }
 
@@ -77,7 +77,7 @@ class PhaseSettingsTest {
         int[] program = {3, 52, 1001, 52, -5, 52, 3, 53, 1, 52, 56, 54, 1007, 54, 5, 55, 1005, 55, 26, 1001, 54,
                 -5, 54, 1105, 1, 12, 1, 53, 54, 53, 1008, 54, 0, 55, 1001, 55, 1, 55, 2, 53, 55, 53, 4,
                 53, 1001, 56, -1, 56, 1005, 56, 6, 99, 0, 0, 0, 0, 10};
-        int output = PhaseSettings.calculateFeedbackThrusterOutput(program, List.of(9, 7, 8, 5, 6));
+        long output = PhaseSettings.calculateFeedbackThrusterOutput(program, List.of(9, 7, 8, 5, 6));
         assertThat(output).isEqualTo(18216);
     }
 
@@ -86,7 +86,7 @@ class PhaseSettingsTest {
         int[] program = {3, 52, 1001, 52, -5, 52, 3, 53, 1, 52, 56, 54, 1007, 54, 5, 55, 1005, 55, 26, 1001, 54,
                 -5, 54, 1105, 1, 12, 1, 53, 54, 53, 1008, 54, 0, 55, 1001, 55, 1, 55, 2, 53, 55, 53, 4,
                 53, 1001, 56, -1, 56, 1005, 56, 6, 99, 0, 0, 0, 0, 10};
-        int output = PhaseSettings.calculateFeedbackThrusterMaxOutput(program);
+        long output = PhaseSettings.calculateFeedbackThrusterMaxOutput(program);
         assertThat(output).isEqualTo(18216);
     }
 
@@ -119,9 +119,9 @@ class PhaseSettingsTest {
     @Test
     void assignment() {
         int[] program = readProgramInput("day7/input.txt");
-        int result1 = PhaseSettings.calculateMaxThrusterOutput(program);
+        long result1 = PhaseSettings.calculateMaxThrusterOutput(program);
         System.out.println("result = " + result1);
-        int result2 = PhaseSettings.calculateFeedbackThrusterMaxOutput(program.clone());
+        long result2 = PhaseSettings.calculateFeedbackThrusterMaxOutput(program.clone());
         System.out.println("result2 = " + result2);
     }
 }
